removed dead pin from binaryCounter.
git-svn-id: https://ktechlab.svn.sourceforge.net/svnroot/ktechlab/trunk@367 76876f85-f02d-0410-881d-fd707285c4ba

+++ src/electronics/components/binarycounter.h
@@ -50,8 +50,8 @@ protected:
 
 	unsigned m_numBits;
 	bool b_triggerHigh;
-	bool b_oldIn;
-	long m_value;
+	bool b_oldIn;  // <<< maybe we should inherit "clocked logic" or something... 
+	unsigned m_value;
 	bool m_bDoneLogicIn;
 };
